MSHACK animation integrated (#1)
* Easier reinstall

* Add details to installation process in Readme

* Change interpreter logic (now uses timestamps)

* WIP: MSHACK2022 animation

* Working version of the MSHACK 2021 animation
