+++ leezenflow_base.py
@@ -16,56 +16,32 @@ from smoother import HoersterTorSmoother
 from simulations import Simulation
 
 class LeezenflowBase(object):
-    def __init__(self, *args, **kwargs):
-        
-        config = configparser.ConfigParser()
-        config.read('config.ini')
+    def __init__(self, command_line_args):
+        try:    
+            config = configparser.ConfigParser()
+            config.read('config.ini')
 
-        self.mqtt_server_ip = config['mqtt']['server_ip']
-        self.mqtt_server_port = int(config['mqtt']['server_port'])
-        self.mqtt_topic =  config['mqtt']['topic']
-        self.mqtt_client_name = config['mqtt']['client_name']
-        self.mqtt_use_auth = config['mqtt']['use_auth']
-        self.mqtt_client_user_name = config['mqtt']['client_user_name']
-        self.mqtt_client_pw = config['mqtt']['client_pw']
+            self.mqtt_server_ip = config['mqtt']['server_ip']
+            self.mqtt_server_port = int(config['mqtt']['server_port'])
+            self.mqtt_topic =  config['mqtt']['topic']
+            self.mqtt_client_name = config['mqtt']['client_name']
+            self.mqtt_use_auth = config['mqtt']['use_auth']
+            self.mqtt_client_user_name = config['mqtt']['client_user_name']
+            self.mqtt_client_pw = config['mqtt']['client_pw']
+        except KeyError:
+            print("Configuration file for MQTT missing...")
+
+        self.args = command_line_args
             
-        # Use Dict such that data can be written at the same  time
+        # Initialize status dictionary that is shared by the interpretation and animation logic
         self.shared_data = {
             "current_phase" : "awaiting_message",
-            "remaining_time" : 30
+            "current_timestamp" : 0,
+            "change_timestamp" : 15, 
         }
-
-        self.parser = argparse.ArgumentParser()
-        self.parser.add_argument("-r", "--led-rows", action="store", help="Display rows. 16 for 16x32, 32 for 32x32. Default: 32", default=32, type=int)
-        self.parser.add_argument("--led-cols", action="store", help="Panel columns. Typically 32 or 64. (Default: 32)", default=96, type=int)
-        self.parser.add_argument("-c", "--led-chain", action="store", help="Daisy-chained boards. Default: 1.", default=1, type=int)
-        self.parser.add_argument("-P", "--led-parallel", action="store", help="For Plus-models or RPi2: parallel chains. 1..3. Default: 1", default=1, type=int)
-        self.parser.add_argument("-p", "--led-pwm-bits", action="store", help="Bits used for PWM. Something between 1..11. Default: 11", default=11, type=int)
-        self.parser.add_argument("-b", "--led-brightness", action="store", help="Sets brightness level. Default: 100. Range: 1..100", default=100, type=int)
-        self.parser.add_argument("-m", "--led-gpio-mapping", help="Hardware Mapping: regular, adafruit-hat, adafruit-hat-pwm" , choices=['regular', 'adafruit-hat', 'adafruit-hat-pwm'], default='adafruit-hat-pwm',type=str)
-        self.parser.add_argument("--led-scan-mode", action="store", help="Progressive or interlaced scan. 0 Progressive, 1 Interlaced (default)", default=1, choices=range(2), type=int)
-        self.parser.add_argument("--led-pwm-lsb-nanoseconds", action="store", help="Base time-unit for the on-time in the lowest significant bit in nanoseconds. Default: 130", default=130, type=int)
-        self.parser.add_argument("--led-show-refresh", action="store_true", help="Shows the current refresh rate of the LED panel")
-        self.parser.add_argument("--led-slowdown-gpio", action="store", help="Slow down writing to GPIO. Range: 0..4. Default: 1", default=3, type=int) #3 best while testing at Swarco; 2 for local mqtt test
-        self.parser.add_argument("--led-no-hardware-pulse", action="store", help="Don't use hardware pin-pulse generation")
-        self.parser.add_argument("--led-rgb-sequence", action="store", help="Switch if your matrix has led colors swapped. Default: RGB", default="RGB", type=str)
-        self.parser.add_argument("--led-pixel-mapper", action="store", help="Apply pixel mappers. e.g \"Rotate:90\"", default="Rotate:0", type=str)
-        self.parser.add_argument("--led-row-addr-type", action="store", help="0 = default; 1=AB-addressed panels; 2=row direct; 3=ABC-addressed panels; 4 = ABC Shift + DE direct", default=0, type=int, choices=[0,1,2,3,4])
-        self.parser.add_argument("--led-multiplexing", action="store", help="Multiplexing type: 0=direct; 1=strip; 2=checker; 3=spiral; 4=ZStripe; 5=ZnMirrorZStripe; 6=coreman; 7=Kaler2Scan; 8=ZStripeUneven... (Default: 0)", default=0, type=int)
-        self.parser.add_argument("--led-panel-type", action="store", help="Needed to initialize special panels. Supported: 'FM6126A'", default="", type=str)
-
-        self.parser.add_argument("--test", action="store", help="Use test dataset. Default: -1; no test", default=-1, type=int)
-        self.parser.add_argument("--logging", action="store", help="1=Log all messages. Default:0 ", default=0, type=int)
 
     def mqtt_client(self,_,run_event):
-        self.shared_data = {
-            "current_phase" : "awaiting_message"
-        }
-        print(self.shared_data)
-
-        interpreter = Interpreter()
-        smoother = HoersterTorSmoother()
-
+        
         def on_connect(client, userdata, flags, rc):
             print("mqtt topic: " + str(self.mqtt_topic))
             print("Connected with result code "+str(rc),flush=True)
@@ -73,6 +49,9 @@ class LeezenflowBase(object):
             # Subscribing in on_connect() means that if we lose the connection and
             # reconnect then subscriptions will be renewed.
             client.subscribe(self.mqtt_topic)
+
+        interpreter = Interpreter()
+        #smoother = HoersterTorSmoother()
 
         def on_message(client, userdata, msg):
             #print(msg.topic+" "+str(msg.payload, "utf-8"),flush=True)
@@ -102,8 +81,6 @@ class LeezenflowBase(object):
 
     def process(self):
 
-        self.args = self.parser.parse_args()
-
         log_name = 'log/all.log'
         if self.args.logging == 1:
             logging.basicConfig(filename=log_name, format='%(asctime)s %(message)s', datefmt='%m/%d/%Y %I:%M:%S %p',level=logging.INFO)
@@ -119,7 +96,6 @@ class LeezenflowBase(object):
         log.addHandler(handler)
 
         options = RGBMatrixOptions()
-
         if self.args.led_gpio_mapping != None:
           options.hardware_mapping = self.args.led_gpio_mapping
         options.rows = self.args.led_rows
@@ -134,10 +110,8 @@ class LeezenflowBase(object):
         options.led_rgb_sequence = self.args.led_rgb_sequence
         options.pixel_mapper_config = self.args.led_pixel_mapper
         options.panel_type = self.args.led_panel_type
-
         if self.args.led_show_refresh:
           options.show_refresh_rate = 1
-
         if self.args.led_slowdown_gpio != None:
             options.gpio_slowdown = self.args.led_slowdown_gpio
         if self.args.led_no_hardware_pulse:
@@ -162,15 +136,16 @@ class LeezenflowBase(object):
         elif self.args.test == 1:
             t2 = threading.Thread(target = Simulation.mqtt_client_simulation, args = (self,None,run_event))
         elif self.args.test == 2:
-            t2 = threading.Thread(target = Simulation.phase_update_simulation, args = (self,"red",run_event))
-        elif self.args.test == 3:
-            t2 = threading.Thread(target = Simulation.phase_switch_simulation, args = (self,6,run_event))
-        elif self.args.test == 4:
-            t2 = threading.Thread(target = Simulation.phase_fast_simulation, args = (self,"green",run_event))
-        elif self.args.test == 5:
             t2 = threading.Thread(target = Simulation.mqtt_client_simulation_dataframe, args = (self,None,run_event))
-        elif self.args.test == 6:
+        elif self.args.test == 3:
+            t2 = threading.Thread(target = Simulation.phase_switch_simulation, args = (self,15,run_event))
+        elif self.args.test == 4:
             t2 = threading.Thread(target = Simulation.phase_switch_simulation, args = (self,30,run_event))            
+        elif self.args.test == 5:
+            t2 = threading.Thread(target = Simulation.phase_update_simulation, args = (self,"red",run_event))
+        elif self.args.test == 6:
+            print("A")
+            t2 = threading.Thread(target = Simulation.phase_fast_simulation, args = (self,"green",run_event))
         else:
             raise Exception("Invalid test dataset argument.")
